Create top bar with a page logo and nav bar
Add a header element and nest a div inside of it to create top bar. Set
dark background color to it. Set display of top-bar to flex to make it
responsive. Add page logo and nav bar inside it. Make nav bar a flex
container to make it responsive. Add some links inside nav bar.